mitigate curl issue observed in macOS builds
macOS installs started failing out of the blue this week for no clear
reason. I think travis-ci overrides curl to make it a little more
failure tolerant, and I'm guessing that something in that process can
cause `source <(curl blah)` to run on an empty FD.

To work around this, I'm breaking it out into two separate steps. I'm
also adding one linux and one macOS smoke-test for this that uses the
same process, so that hopefully it'll break here instead of just
downstream if there's still trouble. (I'm not sure if it makes sense
to use the curl process for *all* of the tests, since there are a fair
number of runs, and curling for each probably increases the odds of
transient failures that must be manually re-run may be reasonable?)